requirements.txt
PyInquirer
pycryptodome
